test: Reorganise files. Tour does not show up.
