+++ virotour/lib/src/app.dart
@@ -1,15 +1,13 @@
 import 'package:flutter/material.dart';
 import 'package:virotour/src/tour/tour_list_view.dart';
 
-import './tour/tour_list_view.dart';
-
 class App extends StatelessWidget {
-  const App({Key? key}) : super(key: key);
+  const App({super.key});
 
   @override
   Widget build(BuildContext context) {
     return MaterialApp(
-      title: 'WebViewX Example App',
+      title: 'ViroTour',
       theme: ThemeData(
         primarySwatch: Colors.blue,
       ),
