+++ virotour/lib/src/tour/tour.dart
@@ -1,0 +1,6 @@
+// A placeholder class that represents an entity or model.
+class Tour {
+  const Tour(this.id);
+
+  final int id;
+}
