+++ virotour/lib/src/tour/tour_list_view.dart
@@ -1,178 +1,68 @@
-import 'dart:math';
-
 import 'package:flutter/material.dart';
-import 'package:webviewx/webviewx.dart';
+import 'package:virotour/src/settings/settings_view.dart';
+import 'package:virotour/src/tour/tour.dart';
+import 'package:virotour/src/tour/tour_details_view.dart';
 
-import '../helpers/helpers.dart';
-
-class TourListView extends StatefulWidget {
+class TourListView extends StatelessWidget {
   const TourListView({
-    Key? key,
-  }) : super(key: key);
+    super.key,
+    this.items = const [Tour(1), Tour(2), Tour(3)],
+  });
 
-  @override
-  _TourListViewState createState() => _TourListViewState();
-}
+  static const routeName = '/';
 
-class _TourListViewState extends State<TourListView> {
-  late WebViewXController webviewController;
-
-  Size get screenSize => MediaQuery.of(context).size;
-
-  @override
-  void dispose() {
-    webviewController.dispose();
-    super.dispose();
-  }
+  final List<Tour> items;
 
   @override
   Widget build(BuildContext context) {
     return Scaffold(
       appBar: AppBar(
         title: const Text('Tours'),
-      ),
-      body: Center(
-        child: Container(
-          padding: const EdgeInsets.all(10.0),
-          child: Column(
-            children: <Widget>[
-              buildSpace(direction: Axis.vertical, amount: 10.0, flex: false),
-              Container(
-                padding: const EdgeInsets.only(bottom: 20.0),
-              ),
-              buildSpace(direction: Axis.vertical, amount: 10.0, flex: false),
-              Container(
-                decoration: BoxDecoration(
-                  border: Border.all(width: 0.2),
-                ),
-                child: _buildWebViewX(),
-              ),
-              Expanded(
-                child: Scrollbar(
-                  isAlwaysShown: true,
-                  child: SizedBox(
-                    width: min(screenSize.width * 0.8, 512),
-                    child: ListView(
-                      children: _buildButtons(),
-                    ),
-                  ),
-                ),
-              ),
-            ],
+        actions: [
+          IconButton(
+            icon: const Icon(Icons.settings),
+            onPressed: () {
+              // Navigate to the settings page. If the user leaves and returns
+              // to the app after it has been killed while running in the
+              // background, the navigation stack is restored.
+              Navigator.restorablePushNamed(context, SettingsView.routeName);
+            },
           ),
-        ),
+        ],
       ),
-    );
-  }
 
-  Widget _buildWebViewX() {
-    return WebViewX(
-      key: const ValueKey('webviewx'),
-      initialSourceType: SourceType.html,
-      height: screenSize.height / 2,
-      width: min(screenSize.width * 0.8, 1024),
-      onWebViewCreated: (controller) => webviewController = controller,
-      onPageStarted: (src) =>
-          debugPrint('A new page has started loading: $src\n'),
-      onPageFinished: (src) =>
-          debugPrint('The page has finished loading: $src\n'),
-      jsContent: const {
-        EmbeddedJsContent(
-          js: "function testPlatformIndependentMethod() { console.log('Hi from JS') }",
-        ),
-        EmbeddedJsContent(
-          webJs:
-              "function testPlatformSpecificMethod(msg) { TestDartCallback('Web callback says: ' + msg) }",
-          mobileJs:
-              "function testPlatformSpecificMethod(msg) { TestDartCallback.postMessage('Mobile callback says: ' + msg) }",
-        ),
-      },
-      dartCallBacks: {
-        DartCallback(
-          name: 'TestDartCallback',
-          callBack: (msg) => showSnackBar(msg.toString(), context),
-        )
-      },
-      webSpecificParams: const WebSpecificParams(
-        printDebugInfo: true,
-      ),
-      mobileSpecificParams: const MobileSpecificParams(
-        androidEnableHybridComposition: true,
-      ),
-      navigationDelegate: (navigation) {
-        debugPrint(navigation.content.sourceType.toString());
-        return NavigationDecision.navigate;
-      },
-    );
-  }
+      // To work with lists that may contain a large number of items, it’s best
+      // to use the ListView.builder constructor.
+      //
+      // In contrast to the default ListView constructor, which requires
+      // building all Widgets up front, the ListView.builder constructor lazily
+      // builds Widgets as they’re scrolled into view.
+      body: ListView.builder(
+        // Providing a restorationId allows the ListView to restore the
+        // scroll position when a user leaves and returns to the app after it
+        // has been killed while running in the background.
+        restorationId: 'tourListView',
+        itemCount: items.length,
+        itemBuilder: (BuildContext context, int index) {
+          final item = items[index];
 
-  void _setUrl() {
-    webviewController.loadContent(
-      'https://cdn.pannellum.org/2.5/pannellum.htm#panorama=https%3A//i.imgur.com/O9CBhdM.jpg&autoLoad=true',
-      SourceType.url,
-    );
-  }
-
-  void _setUrlBypass() {
-    webviewController.loadContent(
-      'https://news.ycombinator.com/',
-      SourceType.urlBypass,
-    );
-  }
-
-  void _setHtmlFromAssets() {
-    webviewController.loadContent(
-      'assets/test.html',
-      SourceType.html,
-      fromAssets: true,
-    );
-  }
-
-  void _toggleIgnore() {
-    final ignoring = webviewController.ignoresAllGestures;
-    webviewController.setIgnoreAllGestures(!ignoring);
-    showSnackBar('Ignore events = ${!ignoring}', context);
-  }
-
-  Future<void> _getWebviewContent() async {
-    try {
-      final content = await webviewController.getContent();
-      showAlertDialog(content.source, context);
-    } catch (e) {
-      showAlertDialog('Failed to execute this task.', context);
-    }
-  }
-
-  Widget buildSpace({
-    Axis direction = Axis.horizontal,
-    double amount = 0.2,
-    bool flex = true,
-  }) {
-    return flex
-        ? Flexible(
-            child: FractionallySizedBox(
-              widthFactor: direction == Axis.horizontal ? amount : null,
-              heightFactor: direction == Axis.vertical ? amount : null,
+          return ListTile(
+            title: Text('Tour ${item.id}'),
+            leading: const CircleAvatar(
+              foregroundImage: AssetImage('assets/images/virotour_logo.png'),
             ),
-          )
-        : SizedBox(
-            width: direction == Axis.horizontal ? amount : null,
-            height: direction == Axis.vertical ? amount : null,
+            onTap: () {
+              // Navigate to the details page. If the user leaves and returns to
+              // the app after it has been killed while running in the
+              // background, the navigation stack is restored.
+              Navigator.restorablePushNamed(
+                context,
+                TourDetailsView.routeName,
+              );
+            },
           );
-  }
-
-  List<Widget> _buildButtons() {
-    return [
-      buildSpace(direction: Axis.vertical, flex: false, amount: 20.0),
-      createButton(
-        text: 'Viet\'s Pannellum Tour',
-        onTap: _setUrl,
+        },
       ),
-      buildSpace(direction: Axis.vertical, flex: false, amount: 20.0),
-      createButton(
-        text: 'https://news.ycombinator.com/',
-        onTap: _setUrlBypass,
-      ),
-    ];
+    );
   }
 }
